refactor(Sort): Sort enum의 sql문 변환책임 제거, 디렉터리 변경
