+++ src/main/java/roomescape/theme/controller/StringToSortColumnConverter.java
@@ -3,6 +3,7 @@ package roomescape.theme.controller;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Component;
+import roomescape.theme.domain.SortColumn;
 
 @Component
 public class StringToSortColumnConverter implements Converter<String, SortColumn> {
