+++ src/main/java/roomescape/theme/controller/StringToSortOrderConverter.java
@@ -3,6 +3,7 @@ package roomescape.theme.controller;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Component;
+import roomescape.theme.domain.SortOrder;
 
 @Component
 public class StringToSortOrderConverter implements Converter<String, SortOrder> {
